Extract build actions to a sub folder
