+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/build_actions/BuildAction.java
@@ -1,0 +1,5 @@
+package brut.androlib.build_actions;
+
+public abstract class BuildAction {
+
+}
