+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/build_actions/BuildLibraryAction.java
@@ -1,0 +1,15 @@
+package brut.androlib.build_actions;
+
+import brut.androlib.build_actions.BuildAction;
+
+public final class BuildLibraryAction extends BuildAction {
+    private final String mFolder;
+
+    public BuildLibraryAction(String folder) {
+        mFolder = folder;
+    }
+
+    public String getFolder() {
+        return mFolder;
+    }
+}
