+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/build_actions/CopyOriginalFilesAction.java
@@ -1,0 +1,6 @@
+package brut.androlib.build_actions;
+
+import brut.androlib.build_actions.BuildAction;
+
+public final class CopyOriginalFilesAction extends BuildAction {
+}
